+++ sales_management_mobile_app/WebAPI/Models/User.cs
@@ -15,8 +15,8 @@ namespace WebAPI.Models
         public string Phone { get; set; }
         public string Address { get; set; }
         public string ManagerId { get; set; }
-        public decimal KpiValue { get; set; }
-        public decimal ActualKpi { get; set; }
+        public int KpiValue { get; set; }
+        public int ActualKpi { get; set; }
        
         public bool IsActive { get; set; }
